Refactor: User, Auth and Admin.
Split package for each entity and refactor all their classes to their
respective package in the new architecture.
