+++ demo/src/main/java/com/project/demo/ApplicationConfiguration.java
@@ -1,6 +1,6 @@
-package com.project.demo.entity.auth;
+package com.project.demo;
 
-import com.project.demo.entity.user.UserRepository;
+import com.project.demo.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
