+++ demo/src/main/java/com/project/demo/entity/User.java
@@ -1,5 +1,4 @@
-package com.project.demo.entity.user;
-import com.project.demo.entity.rol.Role;
+package com.project.demo.entity;
 import jakarta.persistence.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
