+++ demo/src/main/java/com/project/demo/logic/RoleSeeder.java
@@ -1,5 +1,8 @@
-package com.project.demo.entity.rol;
+package com.project.demo.logic;
 
+import com.project.demo.entity.Role;
+import com.project.demo.entity.RoleEnum;
+import com.project.demo.repository.RoleRepository;
 import org.springframework.context.ApplicationListener;
 import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.stereotype.Component;
