+++ demo/src/main/java/com/project/demo/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.project.demo.entity.user;
+package com.project.demo.repository;
 
+import com.project.demo.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
